优化分片下载，小于最小分片数的问题

+++ baidu-netdisk-cp/src/main/java/com/zhonghang/baidu/netdisk/cp/service/SuperFileService.java
@@ -170,7 +170,9 @@ public class SuperFileService {
             Map<String, String> param = HttpUtil.decodeParamMap(request.getUri().toString(),"utf-8");
             requestUtil.requestFile(param,files.get(i),request,stsInfo);
             log.debug("正在上传分片文件{}",  i);
-            files.get(i).delete();
+            if(files.size() != 1){
+                files.get(i).delete();
+            }
         }
     }
